WIP: refactor utils
